requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pyvista==0.46.3
pyvistaqt==0.11.3
QtPy==2.4.3
requests==2.32.5
scooby==0.10.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
vtk==9.5.1

# Additions below for web app and visualization

# --- Web framework ---
Flask>=3.0.0
Werkzeug>=3.0.0

# --- Visualization ---
plotly>=5.20.0
kaleido>=0.2.1          # for saving Plotly figures as static images

# --- Image and data handling ---
Pillow>=10.0.0
numpy>=1.26.0

# --- Whole-slide image (VSI) reading ---
# libopenslide
# openslide-python>=1.3.1   # Fully open-source alternative to BioFormats/Javabridge

bioio>=0.10.0
bioio-ome-tiff>=0.10.0


# --- Bio-Formats / VSI file handling (legacy, to be removed) ---
# python-javabridge>=4.0.3
# python-bioformats>=4.0.0

# --- Optional (for legacy reference or transition scripts) ---
PyQt5>=5.15.0
pyvista>=0.43.0
pyvistaqt>=0.9.0
